+++ src/aws-api-mcp-server/CHANGELOG.md
@@ -12,11 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Updated default AWS API connect and read timeout (#1876)
 - Upgrade AWS CLI to v1.43.10 (#1891)
 - Support `aws login` (#1873)
+- AWS CLI operations 'help' command support (#1858)
 
 ## [1.1.8] - 2025-11-28
 
 ### Fixed
-
 - Origin header parsing (#1851)
 
 ## [1.1.7] - 2025-11-20